refactor: remove deprecated academics v1 api
